src/requirements.txt
python-telegram-bot==21.6
psycopg[binary]==3.2.1
python-dotenv==1.0.1
